src/main/java/io/cherrytechnologies/springrestserge/ui/controllers/HealthController.java
package io.cherrytechnologies.springrestserge.ui.controllers;

import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/v1/api")
public class HealthController {

    @GetMapping(
            path = "/status",
            produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
    public ResponseEntity<String> checkStatus() {
        return ResponseEntity
                .ok("Server is up");
    }
}
